FileStruct Change
